feat: ordena aldatzen du
