+++ front/src/pages/index.tsx
@@ -5,48 +5,28 @@ import { GlobalStyles } from '../ui/GlobalStyles';
 import { Helmet } from 'react-helmet';
 import { Gainburua } from '../components/Gainburua';
 import styled from 'styled-components';
-import { colors, font, fontWeight, media, size } from '../ui/theme';
+import { colors, font, fontWeight, size } from '../ui/theme';
 import { Container } from '../components/Container';
 import { rem } from 'polished';
+import Gezia from '../assets/gezia.svg';
 import ReactMarkdown from 'react-markdown';
 import { Oina } from '../components/Oina';
-import { SFEgunaEdukia } from '../components/SFEgunaEdukia';
-
-export interface SFEguna {
-  id: string;
-  eguna: string;
-  ekitaldi_nagusia: {
-    id: string;
-    hitzordua: string;
-    izenburua: string;
-  };
-}
 
 interface DataProps {
-  strapiHasiera: {
+  strapiLabaGara: {
     deskribapena: string;
     izenburua: string;
-    edukia?: string;
-    sf_egunak?: SFEguna[];
+    edukia: string;
   };
 }
 
 const IndexPage: React.VFC<PageProps> = () => {
-  const { strapiHasiera } = useStaticQuery<DataProps>(graphql`
+  const { strapiLabaGara } = useStaticQuery<DataProps>(graphql`
     {
-      strapiHasiera {
+      strapiLabaGara {
         deskribapena
         izenburua
         edukia
-        sf_egunak {
-          eguna
-          id
-          ekitaldi_nagusia {
-            id
-            hitzordua
-            izenburua
-          }
-        }
       }
     }
   `);
@@ -54,38 +34,32 @@ const IndexPage: React.VFC<PageProps> = () => {
   return (
     <>
       <Helmet
-        title={`${strapiHasiera.izenburua} | Laba`}
+        title={`${strapiLabaGara.izenburua} | Laba`}
         htmlAttributes={{ lang: 'eu' }}
       >
-        <meta name="description" content={strapiHasiera.deskribapena} />
+        <meta name="description" content={strapiLabaGara.deskribapena} />
       </Helmet>
 
       <GlobalStyles />
 
       <Gainburua
         atala="hasiera"
-        izenburua={strapiHasiera.izenburua}
-        deskribapena={strapiHasiera.deskribapena}
+        izenburua={strapiLabaGara.izenburua}
+        deskribapena={strapiLabaGara.deskribapena}
         onClick={() => {
-          window.location.replace('/#edukia');
+          window.location.replace('/#labazkidetza');
         }}
       />
 
-      <ContentWrapper id="edukia">
+      <ContentWrapper id="labazkidetza">
         <Container>
-          {strapiHasiera.edukia && (
-            <Deskribapena>{strapiHasiera.edukia}</Deskribapena>
-          )}
+          <Deskribapena>{strapiLabaGara.edukia}</Deskribapena>
 
-          {strapiHasiera.sf_egunak && strapiHasiera.sf_egunak.length > 0 && (
-            <SFEgunZerrenda>
-              {strapiHasiera.sf_egunak.map(sfeguna => (
-                <SFEgunaElementua key={sfeguna.id}>
-                  <SFEgunaEdukia sfeguna={sfeguna} />
-                </SFEgunaElementua>
-              ))}
-            </SFEgunZerrenda>
-          )}
+          <IzanLabaWrapper>
+            <Botoia href="https://forms.gle/wV41CVkRX1JtHsAu9">
+              <GeziaLogo /> Labazkidetu.
+            </Botoia>
+          </IzanLabaWrapper>
         </Container>
       </ContentWrapper>
 
@@ -93,32 +67,6 @@ const IndexPage: React.VFC<PageProps> = () => {
     </>
   );
 };
-
-const SFEgunaElementua = styled.li`
-  padding: ${rem(size.small)};
-
-  width: 100%;
-  background-color: ${colors.beltza};
-
-  ${media.tablet`
-    aspect-ratio: 1/1;
-    margin-bottom: 0;
-  `}
-`;
-
-const SFEgunZerrenda = styled.ul`
-  display: grid;
-  grid-column-gap: ${rem(size.mini)};
-  grid-row-gap: ${rem(size.mini)};
-
-  ${media.tablet`
-    grid-template-columns: repeat(2, 1fr);
-  `}
-
-  ${media.desktop`
-    grid-template-columns: repeat(3, 1fr);
-  `}
-`;
 
 const Deskribapena = styled(ReactMarkdown)`
   margin-bottom: ${rem(size.large)};
@@ -132,6 +80,31 @@ const ContentWrapper = styled.div`
   padding: ${size.huge}px 0;
   font-weight: ${fontWeight.regular};
   ${font.large()};
+`;
+
+const Botoia = styled.a`
+  border: 2px solid ${colors.beltza};
+  ${font.large()};
+  padding: ${rem(size.base)};
+  background: ${colors.zuria};
+  font-weight: ${fontWeight.bold};
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+
+  transition: color 500ms ease-out 100ms;
+  &:hover {
+    color: ${colors.morea};
+  }
+`;
+
+const GeziaLogo = styled(Gezia)`
+  margin-right: ${rem(size.tiny)};
+`;
+
+const IzanLabaWrapper = styled.div`
+  display: flex;
+  justify-content: flex-end;
 `;
 
 export default IndexPage;
